+++ src/youtube_automation/utils/comments/__init__.py
@@ -1,0 +1,17 @@
+"""自チャンネルのコメント自動返信モジュール群."""
+
+from youtube_automation.utils.comments.fetcher import fetch_top_level_comments
+from youtube_automation.utils.comments.history import ReplyHistory
+from youtube_automation.utils.comments.replier import CommentReplier, ReplyPlan
+from youtube_automation.utils.comments.rule_engine import RuleEngine, RuleMatch
+from youtube_automation.utils.comments.template import render_template
+
+__all__ = [
+    "CommentReplier",
+    "ReplyHistory",
+    "ReplyPlan",
+    "RuleEngine",
+    "RuleMatch",
+    "fetch_top_level_comments",
+    "render_template",
+]